Implemented data recovery from Google Drive backup

+++ app/src/main/java/voxtric/com/diabetescontrol/utilities/GoogleDriveInterface.java
@@ -6,6 +6,7 @@ import android.util.Log;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.api.client.extensions.android.http.AndroidHttp;
 import com.google.api.client.googleapis.extensions.android.gms.auth.GoogleAccountCredential;
+import com.google.api.client.googleapis.media.MediaHttpDownloaderProgressListener;
 import com.google.api.client.googleapis.media.MediaHttpUploaderProgressListener;
 import com.google.api.client.http.ByteArrayContent;
 import com.google.api.client.json.gson.GsonFactory;
@@ -14,6 +15,7 @@ import com.google.api.services.drive.DriveScopes;
 import com.google.api.services.drive.model.File;
 import com.google.api.services.drive.model.FileList;
 
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
@@ -42,7 +44,7 @@ public class GoogleDriveInterface
         .build();
   }
 
-  private File navigateFromParentTo(File parent, String nextFolderName) throws IOException
+  private File navigateFromParentTo(File parent, String nextFolderName, boolean createFolderIfMissing) throws IOException
   {
     if (parent == null)
     {
@@ -55,7 +57,7 @@ public class GoogleDriveInterface
       throw new IllegalArgumentException("Parent folder must have File ID.");
     }
 
-    File folder;
+    File folder = null;
     String parentFolderID = parent.getId();
     FileList folderList = m_googleDrive.files().list()
         .setQ(String.format("name = '%s' and '%s' in parents and mimeType = 'application/vnd.google-apps.folder' and trashed = false", nextFolderName, parentFolderID))
@@ -66,11 +68,14 @@ public class GoogleDriveInterface
     List<File> possibleFolders = folderList.getFiles();
     if (possibleFolders == null || possibleFolders.isEmpty())
     {
-      File newFolder = new File();
-      newFolder.setName(nextFolderName);
-      newFolder.setParents(Collections.singletonList(parentFolderID));
-      newFolder.setMimeType("application/vnd.google-apps.folder");
-      folder = m_googleDrive.files().create(newFolder).execute();
+      if (createFolderIfMissing)
+      {
+        File newFolder = new File();
+        newFolder.setName(nextFolderName);
+        newFolder.setParents(Collections.singletonList(parentFolderID));
+        newFolder.setMimeType("application/vnd.google-apps.folder");
+        folder = m_googleDrive.files().create(newFolder).execute();
+      }
     }
     else
     {
@@ -79,18 +84,25 @@ public class GoogleDriveInterface
     return folder;
   }
 
-  private File getParentFolderOfFile(String filePath) throws IOException
+  private File getParentFolderOfFile(String filePath, boolean createFolderIfMissing) throws IOException
   {
     String[] pathComponents = filePath.split("/");
     File fileParent = null;
-    for (int i = 0; i < pathComponents.length - 1; i++)
-    {
-      fileParent = navigateFromParentTo(fileParent, pathComponents[i]);
-    }
-    if (fileParent == null)
+    if (pathComponents.length <= 1)
     {
       fileParent = new File();
       fileParent.setId("root");
+    }
+    else
+    {
+      for (int i = 0; i < pathComponents.length - 1; i++)
+      {
+        fileParent = navigateFromParentTo(fileParent, pathComponents[i], createFolderIfMissing);
+        if (fileParent == null)
+        {
+          break;
+        }
+      }
     }
     return fileParent;
   }
@@ -101,7 +113,7 @@ public class GoogleDriveInterface
     try
     {
       String fileName = new java.io.File(filePath).getName();
-      String parentFolderID = getParentFolderOfFile(filePath).getId();
+      String parentFolderID = getParentFolderOfFile(filePath, true).getId();
 
       File file = new File();
       file.setName(fileName);
@@ -138,22 +150,58 @@ public class GoogleDriveInterface
     return success;
   }
 
+  public synchronized byte[] downloadFile(String filePath, MediaHttpDownloaderProgressListener downloaderProgressListener)
+  {
+    byte[] data = null;
+    try
+    {
+      String fileName = new java.io.File(filePath).getName();
+      String parentFolderID = getParentFolderOfFile(filePath, false).getId();
+      if (parentFolderID != null)
+      {
+        FileList fileList = m_googleDrive.files().list()
+            .setQ(String.format("name = '%s' and '%s' in parents and trashed = false", fileName, parentFolderID))
+            .setSpaces("drive")
+            .setOrderBy("createdTime")
+            .execute();
+        List<File> existingFiles = fileList.getFiles();
+        if (existingFiles != null && !existingFiles.isEmpty())
+        {
+          ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+          Drive.Files.Get fileGet = m_googleDrive.files().get(existingFiles.get(0).getId());
+          fileGet.getMediaHttpDownloader().setProgressListener(downloaderProgressListener);
+          fileGet.executeMediaAndDownloadTo(outputStream);
+          data = outputStream.toByteArray();
+        }
+      }
+    }
+    catch (IOException exception)
+    {
+      Log.e("GoogleDriveInterface", exception.getMessage(), exception);
+    }
+    return data;
+  }
+
   public synchronized boolean deleteFile(String filePath)
   {
     boolean success = false;
     try
     {
       String fileName = new java.io.File(filePath).getName();
-      String parentFolderID = getParentFolderOfFile(filePath).getId();
-      FileList fileList = m_googleDrive.files().list()
-          .setQ(String.format("name = '%s' and '%s' in parents and trashed = false", fileName, parentFolderID))
-          .setSpaces("drive")
-          .setOrderBy("createdTime")
-          .execute();
-      List<File> existingFiles = fileList.getFiles();
-      if (existingFiles != null && !existingFiles.isEmpty())
+      String parentFolderID = getParentFolderOfFile(filePath, false).getId();
+      if (parentFolderID != null)
       {
-        m_googleDrive.files().delete(existingFiles.get(0).getId()).execute();
+        FileList fileList = m_googleDrive.files().list()
+            .setQ(String.format("name = '%s' and '%s' in parents and trashed = false", fileName, parentFolderID))
+            .setSpaces("drive")
+            .setOrderBy("createdTime")
+            .execute();
+        List<File> existingFiles = fileList.getFiles();
+        if (existingFiles != null && !existingFiles.isEmpty())
+        {
+          m_googleDrive.files().delete(existingFiles.get(0).getId()).execute();
+          success = true;
+        }
       }
     }
     catch (IOException exception)
